Refactor savegame codec tests
Group codec tests in a folder.
Extract the codecs fixtures to a separate fixture, so it can be reused.
Fix operation with trusted_scope, it was left empty after tests.
Refactor pytest to use `import pytest` style.
